refactor: Timer CRUD 정의를 제거하고 Focus로 구조 변경
