+++ Projects/TDData/Sources/Repository/FocusRepositoryImpl.swift
@@ -2,9 +2,11 @@ import TDDomain
 
 struct FocusRepositoryImpl: FocusRepository {
     private let service: FocusService
+    private let storage: TimerStorage
     
-    init(service: FocusService) {
+    init(service: FocusService, storage: TimerStorage) {
         self.service = service
+        self.storage = storage
     }
     
     func saveFocus(date: String, targetCount: Int, settingCount: Int, time: Int) async throws {
@@ -21,5 +23,47 @@ struct FocusRepositoryImpl: FocusRepository {
         let focusList = response.convertToFocusList()
         
         return focusList
+    }
+    
+    func fetchTimerSetting() -> TDTimerSetting {
+        guard let dto = storage.fetchTimerSetting() else {
+            return TDTimerSetting()
+        }
+        return dto.convertToTDTimerSetting()
+    }
+    
+    func updateTimerSetting(setting: TDDomain.TDTimerSetting) throws {
+        try storage.updateTimerSetting(
+            TDTimerSettingDTO(
+                maxFocusCount: setting.focusCountLimit,
+                restDuration: setting.restDuration,
+                focusDuration: setting.focusDuration
+            ))
+    }
+    
+    func fetchTimerTheme() -> TDDomain.TDTimerTheme {
+        guard let dto = storage.fetchTheme() else {
+            return .Bboduck
+        }
+        return dto.convertToTDTimerTheme()
+    }
+    
+    func updateTimerTheme(theme: TDDomain.TDTimerTheme) throws {
+        try storage.updateTheme(TDTimerThemeDTO(timerTheme: theme.rawValue))
+    }
+    
+    func fetchFocusCount() -> Int {
+        guard let count = storage.fetchFocusCount() else {
+            return 0
+        }
+        return count
+    }
+    
+    func updateFocusCount(count: Int) throws {
+        try storage.updateFocusCount(count)
+    }
+    
+    func resetFocusCount() throws {
+        try storage.updateFocusCount(0)
     }
 }
