+++ Projects/TDDomain/Sources/UseCase/Timer/FetchTimerSettingUseCase.swift
@@ -3,9 +3,9 @@ public protocol FetchTimerSettingUseCase {
 }
 
 final class FetchTimerSettingUseCaseImpl: FetchTimerSettingUseCase {
-    private let repository: TimerRepository
+    private let repository: FocusRepository
 
-    public init(repository: TimerRepository) {
+    public init(repository: FocusRepository) {
         self.repository = repository
     }
