+++ Projects/TDDomain/Sources/UseCase/Timer/FocusTimerUseCase.swift
@@ -24,13 +24,13 @@ final class FocusTimerUseCaseImpl: FocusTimerUseCase {
     public var isRunning: Bool {
         return timer != nil
     }
-    private let repository: TimerRepository
+    private let repository: FocusRepository
 
     weak var delegate: FocusTimerUseCaseDelegate?
 
     // MARK: - Initializer
 
-    init(repository: TimerRepository) {
+    init(repository: FocusRepository) {
         self.repository = repository
 
     }
